requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=2.0.0 # Apache-2.0
anyjson>=0.3.3 # BSD
six>=1.9.0 # MIT
eventlet!=0.18.3,>=0.18.2 # MIT
iso8601>=0.1.11 # MIT
netaddr!=0.7.16,>=0.7.13 # BSD
oslo.concurrency>=3.8.0 # Apache-2.0
oslo.utils>=3.20.0 # Apache-2.0
PyYAML>=3.10.0 # MIT
jsonschema>=2.0.0,<3.0.0,!=2.5.0  # MIT
